backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-redis==5.4.0
celery==5.4.0
django-celery-beat==2.7.0
gunicorn==23.0.0
minio==7.2.12
django-environ==0.11.2
redis==5.2.1
django-cors-headers==4.3.1
pytest==8.3.2
pytest-django==4.9.0
weasyprint==62.3
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
django-filter==24.3
drf-spectacular==0.27.2
